Minor changes to comms library
Former-commit-id: 2f4361c3cad293b4e0dfa1eca26d6180d81392f7

+++ CC/Sounder/include/comms-lib.h
@@ -56,10 +56,10 @@ public:
     static std::vector<std::complex<float>> FFT(std::vector<std::complex<float>>, int);
     static std::vector<std::complex<float>> IFFT(std::vector<std::complex<float>>, int);
 
-    static double findLTS(std::vector<std::complex<double>> iq);
+    static int findLTS(std::vector<std::complex<double>> iq);
     static std::vector<double> convolve(std::vector<std::complex<double>> const &f, std::vector<std::complex<double>> const &g);
     static std::vector<std::complex<double>> csign(std::vector<std::complex<double>> iq);
-    static void meshgrid(std::vector<double> x_in, std::vector<double> y_in, std::vector<std::vector<double>> &x, std::vector<std::vector<double>> &y);
+    static void meshgrid(std::vector<int> x_in, std::vector<int> y_in, std::vector<std::vector<int>> &x, std::vector<std::vector<int>> &y);
 //private:
 //    static inline float** init_qpsk();
 //    static inline float** init_qam16();
